+++ src/prezoom/model/GObject.java
@@ -38,7 +38,7 @@ public abstract class GObject
      * @param y2 y2 of the object, if applicable
      * @param visible whether visible
      */
-    protected GObject(Double x, Double y, Color col, Boolean filled, Integer lineWidth, Integer width, Integer height, Double x2, Double y2, Boolean visible)
+    protected GObject(Double x, Double y, Color col, Boolean filled, Integer lineWidth, Double width, Double height, Double x2, Double y2, Boolean visible)
     {
         //this.id = id;
 //        this.x = x;
