+++ src/prezoom/model/GOval.java
@@ -9,7 +9,7 @@ import java.awt.geom.Ellipse2D;
  **/
 public class GOval extends GObject
 {
-    public GOval(double x, double y, int w, int h, Color col, Boolean filled, int lineWidth)
+    public GOval(double x, double y, double w, double h, Color col, Boolean filled, int lineWidth)
     {
         super(x, y, col, filled, lineWidth, w, h, null,null,true);
         this.drawShape = new Ellipse2D.Double(x, y, w, h);
@@ -22,7 +22,7 @@ public class GOval extends GObject
     public void draw(Graphics2D g)
     {
         double x = getCurrentAttributes().getX(), y = getCurrentAttributes().getY();
-        int w = getCurrentAttributes().getWidth(), h = getCurrentAttributes().getHeight();
+        double w = getCurrentAttributes().getWidth(), h = getCurrentAttributes().getHeight();
 
         this.drawShape = new Ellipse2D.Double(x, y, w, h);
