+++ src/prezoom/model/GRectangle.java
@@ -9,7 +9,7 @@ import java.awt.geom.Rectangle2D;
  **/
 public class GRectangle extends GObject
 {
-    public GRectangle(double x, double y, int w, int h, Color col, Boolean filled, int lineWidth)
+    public GRectangle(double x, double y, double w, double h, Color col, Boolean filled, int lineWidth)
     {
         super(x, y, col, filled, lineWidth,w,h,null,null,true);
         this.drawShape = new Rectangle2D.Double(x, y, w, h);
@@ -23,7 +23,7 @@ public class GRectangle extends GObject
     public void draw(Graphics2D g)
     {
         double x = getCurrentAttributes().getX(), y = getCurrentAttributes().getY();
-        int w = getCurrentAttributes().getWidth(), h = getCurrentAttributes().getHeight();
+        double w = getCurrentAttributes().getWidth(), h = getCurrentAttributes().getHeight();
 
         this.drawShape = new Rectangle2D.Double(x, y, w, h);
